requirements.txt
Django==1.6.2
django-report-tools==0.2.2
django-toolbelt==0.0.1
gunicorn==18.0
pycha==0.7.0
